feat: instrument with otel
Since we wanted to improve the observability of our application, we will instrument it with OpenTelemetry, so that we can send traces, metrics, and logs.

This commit includes the installation of OpenTelemetry dependencies, the infrastructure as code for the observability stack, and the instrumentation of the classes associated with the Author for sending traces, logs, and metrics.